refactor: remove services/api
